ボタンのコンポーネント追加した
