+++ src/components/SaveButton.vue
@@ -1,0 +1,15 @@
+<template>
+    <v-btn
+        class="ma-2"
+        outlined
+        color=#ea5532
+        elevation="2"
+        rounded
+    >Save</v-btn>
+</template>
+
+<script>
+</script>
+
+<style>
+</style>
